로그인 기능

+++ src/main/java/com/resume_management/resume_management_back/user/dto/LoginResponseDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class LoginResponseDTO {
     
-    private Long user_id;
+    private int user_id;
     private String name;
     private String email          ;
     private String password       ;
